+++ tests/gamemodes/parsers/conftest.py
@@ -1,8 +1,0 @@
-import pytest
-
-from app.gamemodes.parsers.gamemodes_parser import GamemodesParser
-
-
-@pytest.fixture(scope="package")
-def gamemodes_parser() -> GamemodesParser:
-    return GamemodesParser()